re-structuring to improve efficiency
